refactor & fix: add lazy-loading to avante_templates lib init
Sometimes avante_templates takes longer than 1s to load (for example, on
very slow machines, one of my pc load avante for 900ms+), which causes the error
"Make sure to build avante (missing avante_templates)".
This error is incorrect since the library exists but hasn't finished loading.
Now the library is loaded lazily and this error will appear
only when the library is actually missing.
I'm not sure if this is a right solution in this case,
mb better just add templates_load_delay setting to config.